requirements.txt
aioredis==1.3.1
asgiref==3.5.0
async-timeout==4.0.2
attrs==22.1.0
autobahn==23.1.2
Automat==22.10.0
bcrypt==4.0.1
black==22.6.0
blessed==1.19.1
boto3==1.20.54
botocore==1.23.54
cached-property==1.5.2
cement==2.8.2
certifi==2022.9.24
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.3
constantly==15.1.0
cryptography==38.0.1
daphne==3.0.2
dj-database-url==0.5.0
Django==4.0.3
django-cors-headers==3.10.0
django-dotenv==1.4.2
django-heroku==0.3.1
django-model-utils==4.2.0
django-storages==1.13.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
Faker==17.6.0
flake8==5.0.4
flake8-black==0.3.3
future==0.16.0
gunicorn==20.1.0
hiredis==2.2.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jmespath==0.10.0
jsonschema==3.2.0
mccabe==0.7.0
msgpack==1.0.5
mypy-extensions==0.4.3
paramiko==2.11.0
pathspec==0.9.0
Pillow==8.0.1
platformdirs==2.5.2
# psycopg2==2.9.5
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
PyJWT==2.4.0
PyNaCl==1.5.0
pyOpenSSL==23.0.0
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==5.4.1
requests==2.26.0
s3transfer==0.5.2
semantic-version==2.8.5
service-identity==21.1.0
shortuuid==1.0.9
six==1.14.0
sqlparse==0.4.2
termcolor==1.1.0
texttable==1.6.4
tomli==2.0.1
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.3.0
urllib3==1.26.12
wcwidth==0.1.9
websocket-client==0.59.0
whitenoise==6.0.0
zope.interface==5.5.2
